+++ cypress/e2e/EnrollmentEditEventPage/EnrollmentEditEventPageForm/EnrollmentEditEventPageForm.js
@@ -1,4 +1,4 @@
-import { Given, When, Then, defineStep as And } from '@badeball/cypress-cucumber-preprocessor';
+import { defineStep as And, Given, Then, When } from '@badeball/cypress-cucumber-preprocessor';
 import { getCurrentYear } from '../../../support/date';
 
 const changeEnrollmentAndEventsStatus = () => (
@@ -13,7 +13,9 @@ const changeEnrollmentAndEventsStatus = () => (
             const enrollmentToUpdate = {
                 ...enrollment,
                 status: 'ACTIVE',
-                events: [{ ...eventToUpdate, status: 'ACTIVE' }],
+                completedAt: null,
+                completedBy: null,
+                events: [{ ...eventToUpdate, status: 'ACTIVE', completedAt: null, completedBy: null }],
             };
 
             return cy
@@ -209,4 +211,16 @@ Then('the edit button should be disabled', () => {
         .find('[data-test="dhis2-uicore-button"]')
         .eq(1)
         .should('be.disabled');
+});
+
+And('the add event form is displayed', () => {
+    cy.get('[data-test="add-event-enrollment-page-content"]').should('exist');
+});
+
+And('the user is navigated to the enrollment dashboard', () => {
+    cy.get('[data-test="enrollment-overview-page"]').should('exist');
+});
+
+And(/^the view enrollment event form is in (.*) mode$/, (mode) => {
+    cy.get(`[data-test="widget-enrollment-event-${mode}"]`).should('exist');
 });