+++ src/core_modules/capture-core/components/WorkingLists/WorkingListsBase/ContextBuilder/ContextProviders/WorkingListsListViewBuilderContextProvider.component.js
@@ -1,15 +1,18 @@
 // @flow
 import React, { useMemo } from 'react';
-import {
-    ListViewBuilderContext,
-} from '../../workingListsBase.context';
+import { ListViewBuilderContext } from '../../workingListsBase.context';
 import type { Props } from './workingListsListViewBuilderContextProvider.types';
 
 export const WorkingListsListViewBuilderContextProvider = ({
     updating,
     updatingWithDialog,
+    selectedRows,
+    allRowsAreSelected,
+    selectionInProgress,
     dataSource,
-    onSelectListRow,
+    onClickListRow,
+    onRowSelect,
+    onSelectAll,
     onSortList,
     onSetListColumnOrder,
     customRowMenuContents,
@@ -21,13 +24,19 @@ export const WorkingListsListViewBuilderContextProvider = ({
     onChangeRowsPerPage,
     stickyFilters,
     programStageId,
+    bulkActionBarComponent,
     children,
 }: Props) => {
     const listViewBuilderContextData = useMemo(() => ({
         updating,
         updatingWithDialog,
         dataSource,
-        onSelectListRow,
+        selectedRows,
+        allRowsAreSelected,
+        selectionInProgress,
+        onClickListRow,
+        onRowSelect,
+        onSelectAll,
         onSortList,
         onSetListColumnOrder,
         customRowMenuContents,
@@ -39,11 +48,17 @@ export const WorkingListsListViewBuilderContextProvider = ({
         onChangeRowsPerPage,
         stickyFilters,
         programStageId,
+        bulkActionBarComponent,
     }), [
         updating,
         updatingWithDialog,
         dataSource,
-        onSelectListRow,
+        selectedRows,
+        allRowsAreSelected,
+        selectionInProgress,
+        onClickListRow,
+        onRowSelect,
+        onSelectAll,
         onSortList,
         onSetListColumnOrder,
         customRowMenuContents,
@@ -55,6 +70,7 @@ export const WorkingListsListViewBuilderContextProvider = ({
         onChangeRowsPerPage,
         stickyFilters,
         programStageId,
+        bulkActionBarComponent,
     ]);
 
     return (